server/requirements.txt
# requirements.txt - Updated for Flask 3.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
PyJWT==2.8.0  # Using PyJWT instead of flask-jwt-extended
psycopg2-binary==2.9.9
qrcode==7.4.2
Pillow==10.1.0
python-dateutil==2.8.2
pytest==7.4.3
pytest-flask==1.2.0
requests==2.31.0